# contactapi/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "STIS Contact API",
    "description": "API ini menyediakan layanan untuk mengelola dan mengakses informasi kontak yang dimiliki oleh mahasiswa dan dosen di Politeknik Statistika STIS. Layanan ini dirancang untuk mempermudah pencarian informasi kontak dengan menyediakan berbagai endpoint untuk menambah, memperbarui, dan menghapus kontak dari mahasiswa dan dosen. API ini juga mendukung pengelompokan kontak berdasarkan organisasi mahasiswa atau mata kuliah yang diampu oleh dosen serta menawarkan fitur pencarian yang fleksibel. \n\nFitur utama meliputi: \n- Manajemen kontak mahasiswa dan dosen \n- Pengelompokan kontak berdasarkan organisasi atau jurusan \n- Dukungan endpoint untuk CRUD (Create, Read, Update, Delete) pada data kontak \n- Fitur pencarian kontak berdasarkan organisasi atau mata kuliah yang diampu oleh dosen\n\n",
    "contact": {
      "name": "Dodi Firmansyah",
      "url": "https://github.com/dodfrm",
      "email": "222212572@stis.ac.id"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/organizations/{id}": {
      "get": {
        "tags": ["Organization"],
        "summary": "Get organization by ID",
        "operationId": "getOrganizationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Organization not found"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Organization found",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Organization"],
        "summary": "Update an organization",
        "operationId": "updateOrganization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "organizationName": "BULSTIK"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Organization not found"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Organization updated successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Organization"],
        "summary": "Delete an organization",
        "operationId": "deleteOrganization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Organization not found"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Organization deleted successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact by ID",
        "operationId": "getContactById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact not found"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Contact found",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Contacts"],
        "summary": "Update an existing contact",
        "operationId": "updateContact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "fullName": "Dodi Firmansyah",
                  "phone": "081998915680",
                  "email": "222212572@stis.ac.id",
                  "contactType": "MAHASISWA or DOSEN"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact not found"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete a contact",
        "operationId": "deleteContact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact not found"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Contact deleted successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contact-organizations/{id}": {
      "get": {
        "tags": ["Contact Organization"],
        "summary": "Get a contact organization by ID",
        "operationId": "getContactOrganizationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactOrganizationDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact organizations not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact organizations not found"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Contact Organization"],
        "summary": "Update an existing contact organization",
        "operationId": "updateContactOrganization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "contactId": "{id}",
                  "kelas": "3SI1",
                  "organizationId": "{id}",
                  "jabatan": "KETUA/WAKIL_KETUA/BENDAHARA",
                  "periodeJabatan": "2024/2025"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Contact organization updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactOrganizationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Contact organizations not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact organizations not found"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Contact Organization"],
        "summary": "Delete a contact organization by ID",
        "operationId": "deleteContactOrganization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Contact organization deleted successfully"
          },
          "404": {
            "description": "Contact organizations not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact organizations not found"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contact-organizations": {
      "get": {
        "tags": ["Contact Organization"],
        "summary": "Get all contact organizations",
        "operationId": "getAllContactOrganizations",
        "responses": {
          "200": {
            "description": "Successfully retrieved list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactOrganizationDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Contact Organization"],
        "summary": "Create a new contact organization",
        "operationId": "createContactOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "contactId": "{id}",
                  "kelas": "3SI1",
                  "organizationId": "{id}",
                  "jabatan": "KETUA/WAKIL_KETUA/BENDAHARA",
                  "periodeJabatan": "2024/2025"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Contact not found.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactOrganizationDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Contact organization created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactOrganizationDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Subject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "subjectName": {
            "type": "string"
          }
        }
      },
      "ContactSubject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          }
        }
      },
      "ContactOrganization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kelas": {
            "type": "string"
          },
          "jabatan": {
            "type": "string",
            "enum": ["KETUA", "WAKIL_KETUA", "BENDAHARA", "PJ_MATKUL"]
          },
          "periodeJabatan": {
            "type": "string"
          }
        }
      },
...
      "ContactDto": {
        "required": ["fullName", "phone"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "contactType": {
            "type": "string",
            "enum": ["MAHASISWA", "DOSEN"]
          },
          "contactOrganizations": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactOrganization"
            }
          },
          "contactSubjects": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactSubject"
            }
          }
        }
      },
...
      "OrganizationDto": {
        "required": ["organizationName"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "organizationName": {
            "type": "string"
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactDto"
            }
          }
        }
      },
...
      "ContactOrganizationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "contactId": {
            "type": "integer",
            "format": "int64"
          },
          "kelas": {
            "type": "string"
          },
          "organizationId": {
            "type": "integer",
            "format": "int64"
          },
          "jabatan": {
            "type": "string",
            "enum": ["KETUA", "WAKIL_KETUA", "BENDAHARA", "PJ_MATKUL"]
          },
          "periodeJabatan": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Input token sebagai 'Bearer \u003Ctoken\u003E'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}